minor refactor to get the a11y working
